Fixing warnings

+++ source/src/customserver.cpp
@@ -1305,8 +1305,8 @@ namespace CustomServer
                 return;
             }
 
-            int32_t address = request->getParam("address")->value().toInt();
-            int32_t bits = request->getParam("bits")->value().toInt();
+            uint16_t address = request->getParam("address")->value().toInt();
+            uint8_t bits = request->getParam("bits")->value().toInt();
             bool signedData = request->hasParam("signed") ? request->getParam("signed")->value().equals("true") : false;
 
             int32_t value = Ade7953::readRegister(address, bits, signedData);
@@ -1335,8 +1335,8 @@ namespace CustomServer
                     return;
                 }
 
-                int32_t address = doc["address"].as<int32_t>();
-                int32_t bits = doc["bits"].as<int32_t>();
+                uint16_t address = doc["address"].as<int32_t>();
+                uint8_t bits = doc["bits"].as<int32_t>();
                 int32_t value = doc["value"].as<int32_t>();
 
                 Ade7953::writeRegister(address, bits, value);
